pyninja: Python interface to ninja internals
Using SWIG we can provide a Python interface to ninja internals.  This
allows the rapid development of tools that can instrument ninja's build
process in various ways.  At the moment the generated module is usable
as a drop-in replacement for the ninja binary (without the -d and -t
options).  Only APIs that were necessary for this goal were wrapped.
Others will follow in future changes.

Change-Id: I546a9bf1a64de778d93f76abde2c430c3fc3055a