hash based input checking
When a rule specifies "hash_input = 1" in addition to the modification
time of its inputs the hashed content of the inputs are taken into
account.  Hence a target is only dirty if the incoming edges are dirty
content-wise.

The purpose of the change is addressing workflows where input files are
rewritten but their content remains the same such.  This frequently
happens when working with version control systems, for example when
rebasing or switching branches.

For every completed edge the filenames of the outputs and filenames,
mtimes and 32bit Murmur3 hashes for the inputs are stored.

The mtime of a hashed files are stored to so that subsequent checks of
touched but clean files can compare that instead of recomputing the
hash.

The performance impact of the feature when it is disabled is negligible.
When it is enabled rereading the files and computing the hashes mostly
unnoticable - especially when the commands themselves are relatively
expensive, such as compiling C++ source files.

Co-Authored-By: Frank Benkstein <frank.benkstein@sap.com>